LayoutTests/platform/chromium-win/svg/custom/polyline-points-crash-expected.txt

CONSOLE MESSAGE: line 6: Uncaught Error: TYPE_MISMATCH_ERR: DOM Exception 17
